Scope irregular singularization of lice and mice
This irregular rule is being over-applied to things like

- slice -> slouse
- accomplice -> accomplouse

when run through `singularize` or `inflect` with a `1`.

The irregular singularization is limited to _lice_ and _mice_ as far as
I can tell. That is, there aren't longer or compound words ending in
_lice_ or _mice_ that also have an irregular singular form. Hence, I
believe an appropriate solution is to cap the regex for this rule to the
beginning of the line.